+++ src/SeoMaestroException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace SeoMaestro;
+
+use ProcessWire\WireException;
+
+class SeoMaestroException extends WireException
+{
+
+}